Changed source API. Added currencies to CSV to avoid unnecessary API calls

+++ src/main/java/software/ulpgc/moneycalculator/model/Money.java
@@ -1,6 +1,6 @@
 package software.ulpgc.moneycalculator.model;
 
-public record Money(long amount, Currency currency) {
+public record Money(double amount, Currency currency) {
     @Override
     public String toString() {
         return amount + " " + currency;
